Initial commit of AllStarLink's asterisk files needed to compile branch on 3.x+ linux kernels.
DAHDI - contains the 2.10.2 version of the DAHDI driver from the digium repository along with a patch that will make the driver work with AllStarlink (ported from https://github.com/anthcp).  Patch was for previous version of DAHDI driver.  I ported the changes to the 2.10.2 driver. (This driver is currently in use with my own AllStarLink node - 41638)

chan modules - contains the chan_alsaradio.c and config file. This is supposed to allow you to use a USB soundcard and serial port for PTT control with a radio.

etc - contains a script for pulling down the latest/current nodelist.

Makefiles - Makefile.trunk, Makefile.gsm, and Makefile.libpri are additional makefiles pulled from the same repository as the DAHDI driver above.  .trunk replaces the makefile at the top of the trunk folder, .libpri replaces the makefile in the libpri folder, and .gsm replaces the makefile in the codecs/gsm folder.  (NOte:  these makefiles have already been copied to their respective places in the trunk folder)

trunk - contains the modifications to the latest "trunk" (at the time of commit) of the AllStarLink source.  All Makefiles have been replaced and the configure script (./configure and configure.ac) have been changed to ensure that asterisk and all required files compile correctly.